+++ Logger/Entities/People/Student.cs
@@ -2,11 +2,11 @@
 
 public record class Student : Person
 {
-    public float Gpa { get; init; }
+    public decimal Gpa { get; init; }
 
-    public Student(FullName fullName, float gpa) : base(fullName)
+    public Student(FullName fullName, decimal gpa) : base(fullName)
     {
-        if (gpa < 0 || gpa > 4.0) throw new ArgumentOutOfRangeException(nameof(gpa), "GPA must be between 0 and 4.0");
+        if (gpa < 0 || gpa > 4) throw new ArgumentOutOfRangeException(nameof(gpa), "GPA must be between 0 and 4.0");
         Gpa = gpa;
     }
 }
